Panic and log macros

+++ c_version/include/string.h
@@ -5,6 +5,6 @@
 
 void memset(void *va, u8 pattern, int len);
 int memcmp(const u8 *va, const char *string, int len);
-u32 sum_bytes(u8 *va, int count);
+u8 sum_bytes(u8 *va, int count);
 
 #endif  // XV6_STRING_H
